# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TestTask",
        "description": "TestTask OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user-profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user information",
                "description": "Get user information with orders and its information",
                "operationId": "getUserProfile",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "readOnly": "true",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "description": "User unique email address",
                                                    "type": "string",
                                                    "format": "email",
                                                    "readOnly": "true",
                                                    "example": "user@gmail.com"
                                                },
                                                "orders": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "code": {
                                                                "type": "string",
                                                                "example": "AAAAAAAAAAAA"
                                                            },
                                                            "start_time": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2022-06-11 21:36:35"
                                                            },
                                                            "finish_time": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2022-06-11 21:36:35"
                                                            },
                                                            "status_id": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "cancelled"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "example": [
                                                        {
                                                            "id": 1,
                                                            "code": "AAAAAAAAAAAA",
                                                            "start_time": "2022-06-11 21:36:35",
                                                            "finish_time": "2022-06-11 21:36:35",
                                                            "status_id": 3,
                                                            "status": "cancelled"
                                                        }
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "validation errors."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
...
}